allow user to choose between CENNZnet or MetaMask wallets
