+++ components/WalletButton.tsx
@@ -1,56 +1,77 @@
-import React, { useState, useMemo, useCallback } from "react";
+import React, { useMemo } from "react";
 import { css } from "@emotion/react";
 import { useCENNZWallet } from "@/providers/CENNZWalletProvider";
 import WalletModal from "@/components/WalletModal";
 import AccountIdenticon from "@/components/shared/AccountIdenticon";
 import CENNZIconSVG from "@/assets/vectors/cennznet-icon.svg";
 import { Theme } from "@mui/material";
+import WalletSelect from "@/components/WalletSelect";
+import Wallet from "@/components/Wallet";
+import { useWalletProvider } from "@/providers/WalletProvider";
+import { useMetaMaskWallet } from "@/providers/MetaMaskWalletProvider";
+import Jazzicon, { jsNumberForAddress } from "react-jazzicon";
 
-type WalletState = "NotConnected" | "Connecting" | "Connected";
+type WalletState = "Connected" | "NotConnected";
 
-const WalletButton: React.FC<{}> = () => {
-	const [modalOpen, setModalOpen] = useState<boolean>(false);
-	const { selectedAccount, balances, connectWallet } = useCENNZWallet();
+const WalletButton: React.FC = () => {
+	const { walletOpen, setWalletOpen, selectedWallet } = useWalletProvider();
+	const { selectedAccount: CENNZAccount } = useCENNZWallet();
+	const { selectedAccount: metaMaskAccount } = useMetaMaskWallet();
+
 	const walletState = useMemo<WalletState>(() => {
-		if (!selectedAccount) return "NotConnected";
-		if (selectedAccount && !balances?.length) return "Connecting";
-		return "Connected";
-	}, [selectedAccount, balances]);
-
-	const onWalletClick = useCallback(async () => {
-		if (walletState !== "NotConnected") return setModalOpen(true);
-
-		await connectWallet();
-		setModalOpen(true);
-	}, [walletState, connectWallet]);
+		if (!!selectedWallet) return "Connected";
+		if (!CENNZAccount || !metaMaskAccount) return "NotConnected";
+	}, [CENNZAccount, metaMaskAccount, selectedWallet]);
 
 	return (
 		<>
-			<div css={styles.walletButton(modalOpen)} onClick={onWalletClick}>
+			<div
+				css={styles.walletButton(walletOpen)}
+				onClick={() => setWalletOpen(true)}
+			>
 				<div css={styles.walletIcon}>
-					<img
-						src={CENNZIconSVG.src}
-						alt="CENNZnet Logo"
-						css={styles.walletIconImg}
-					/>
+					{(walletState === "NotConnected" || !selectedWallet) && (
+						<img
+							src={CENNZIconSVG.src}
+							alt="CENNZnet Logo"
+							css={styles.walletIconImg}
+						/>
+					)}
 
-					{!!selectedAccount?.address && (
+					{!!CENNZAccount?.address && selectedWallet === "CENNZnet" && (
 						<AccountIdenticon
 							css={styles.walletIconIdenticon}
 							theme="beachball"
 							size={28}
-							value={selectedAccount.address}
+							value={CENNZAccount.address}
+						/>
+					)}
+					{!!metaMaskAccount?.address && selectedWallet === "MetaMask" && (
+						<Jazzicon
+							diameter={28}
+							seed={jsNumberForAddress(metaMaskAccount?.address as string)}
 						/>
 					)}
 				</div>
+
 				<div css={styles.walletState}>
-					{walletState !== "NotConnected" && (
-						<span>{selectedAccount?.meta?.name?.toUpperCase?.()}</span>
+					{walletState === "Connected" && selectedWallet === "CENNZnet" && (
+						<span>{CENNZAccount?.meta?.name?.toUpperCase?.()}</span>
 					)}
-					{walletState === "NotConnected" && <span>CONNECT CENNZnet</span>}
+					{walletState === "Connected" && selectedWallet === "MetaMask" && (
+						<span>
+							{metaMaskAccount?.address
+								.slice(0, 6)
+								.concat("...", metaMaskAccount?.address.slice(-4))}
+						</span>
+					)}
+					{!selectedWallet && <span>CONNECT WALLET</span>}
 				</div>
 			</div>
-			<WalletModal setModalOpen={setModalOpen} modalOpen={modalOpen} />
+			<WalletModal>
+				{!selectedWallet && <WalletSelect />}
+				{selectedWallet && <Wallet />}
+			</WalletModal>
 		</>
 	);
 };
@@ -59,7 +80,7 @@ export default WalletButton;
 
 export const styles = {
 	walletButton:
-		(modalOpen: boolean) =>
+		(walletOpen: boolean) =>
 		({ palette, shadows, transitions }: Theme) =>
 			css`
 				position: absolute;
@@ -70,8 +91,8 @@ export const styles = {
 				height: 48px;
 				display: flex;
 				align-items: center;
-				background-color: ${modalOpen ? palette.primary.default : "#FFFFFF"};
-				color: ${modalOpen ? "#FFFFFF !important" : palette.primary.default};
+				background-color: ${walletOpen ? palette.primary.default : "#FFFFFF"};
+				color: ${walletOpen ? "#FFFFFF !important" : palette.primary.default};
 				transition: background-color ${transitions.duration.short}ms,
 					color ${transitions.duration.short}ms;
 				border-radius: 4px;
